+++ Source/Modules/Catalog/IO/Gapotchenko.FX.IO/PathEx.cs
@@ -1,5 +1,6 @@
 ﻿// Gapotchenko.FX
 // Copyright © Gapotchenko and Contributors
+//
 // Portions © .NET Foundation and its Licensors
 //
 // File introduced by: Oleksiy Gapotchenko